requirements.txt
bcrypt==3.1.6
boto3==1.9.82
botocore==1.12.82
cffi==1.11.5
Django==2.1.2
django-storages==1.7.1
docutils==0.14
jmespath==0.9.3
Pillow==5.4.1
pycparser==2.19
python-dateutil==2.7.5
pytz==2018.9
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
